+++ R/evalOptimizationState.R
@@ -13,6 +13,7 @@ evalOptimizationState = function(learner, task, resampling, measures, par.set, b
   learner2 = learner
   threshold = NULL
   log.fun = control$log.fun
+  err.dumps = list()
 
   if (inherits(control, "TuneControl") || inherits(control, "TuneMultiCritControl")) {
     # set names before trafo
@@ -61,6 +62,7 @@ evalOptimizationState = function(learner, task, resampling, measures, par.set, b
     notna = !is.na(errmsgs)
     if (any(notna))
       errmsg = errmsgs[notna][1L]
+    err.dumps = r$err.dumps
   } else {
     # we still need to define a non-NULL threshold, if tuning it was requested
     if (control$tune.threshold)
@@ -71,7 +73,8 @@ evalOptimizationState = function(learner, task, resampling, measures, par.set, b
   if (show.info)
     log.fun(learner, task, resampling, measures, par.set, control, opt.path, dob, state, y,
       remove.nas, stage = 2L, prev.stage = prev.stage)
-  list(y = y, exec.time = exec.time, errmsg = errmsg, threshold = threshold)
+  list(y = y, exec.time = exec.time, errmsg = errmsg, threshold = threshold,
+      err.dumps = err.dumps)
 }
 
 # evaluates a list of states by calling evalOptimizationState
@@ -95,10 +98,19 @@ evalOptimizationStates = function(learner, task, resampling, measures, par.set, 
       measures = measures, par.set = par.set, bits.to.features = bits.to.features,
       control = control, opt.path = opt.path, show.info = show.info, remove.nas = remove.nas,
       resample.fun = resample.fun))
+
+  on.error.dump = getMlrOption("on.error.dump")
   # add stuff to opt.path
   for (i in seq_len(n)) {
     res = res.list[[i]]
     extra = getTuneThresholdExtra(control, res)
+    # include error dumps if options tell us to. 
+    if (on.error.dump) {
+      if (is.null(extra)) {
+        extra = list()
+      }
+      extra$.dump = res$err.dumps
+    }
     addOptPathEl(opt.path, x = as.list(states[[i]]), y = res$y, exec.time = res$exec.time,
       error.message = res$errmsg, dob = dobs[i], eol = eols[i], check.feasible = TRUE,
       extra = extra)