+++ R/tuneParams.R
@@ -115,7 +115,8 @@ tuneParams = function(learner, task, resampling, measures, par.set, control, sho
   if (cl == "TuneControlMBO" && !identical(resample.fun, resample)) {
     stop("resample.fun must be the default value when mbo tuning is performed.")
   } 
-  opt.path = makeOptPathDFFromMeasures(par.set, measures, include.extra = (control$tune.threshold))
+  need.extra = control$tune.threshold || getMlrOption("on.error.dump")
+  opt.path = makeOptPathDFFromMeasures(par.set, measures, include.extra = need.extra)
   if (show.info) {
     messagef("[Tune] Started tuning learner %s for parameter set:", learner$id)
     message(printToChar(par.set))  # using message() since this can go over the char limit of messagef(), see issue #1528