+++ R/tuneParamsMultiCrit.R
@@ -65,7 +65,7 @@ tuneParamsMultiCrit = function(learner, task, resampling, measures, par.set, con
     TuneMultiCritControlNSGA2 = tuneMultiCritNSGA2,
     stopf("Tuning algorithm for '%s' does not exist!", cl)
   )
-  opt.path = makeOptPathDFFromMeasures(par.set, measures)
+  opt.path = makeOptPathDFFromMeasures(par.set, measures, include.extra = getMlrOption("on.error.dump"))
   if (show.info) {
     messagef("[Tune] Started tuning learner %s for parameter set:", learner$id)
     messagef(printToChar(par.set))